Create a service object to validate user as a member of DevBootcamp's Owners team
I had  begun riting a GitHubUser class where instances would validate themselves as members of the team, but after reading some blog posts I wanted to extract this out to a service object that is only concerned with this aspect of validation.  Give it a github_token and it tells you whether or not the user to whom the token belongs is a member of the DevBootcamp Owner's team.

In order to test communicating with the GitHub API, I used WebMock to stub responses to the various requests being made.