Render a template for new groups page
Clicking to make new groups for a cohort, makes a new grouping object for the cohort and then redirects to the page for creating new groups.  In this commit, the proper template is being rendered.  Not all of the desired behavior for the GroupsController#new action is complete, but doing so will require a number of changes elsewhere in the app, so this intermediate commit will serve as the jumping off point for those changes.